+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Allow declaring functions that take no parameters.
 - Fix readme typo.
 - Correctly instantiate Parameter objects when not passed via the DSL.
+- Raise an exception if the API returns an error code.